chore(test): upgrade all deps
In order to use the latest Next.js features (server actions), we should upgrade
these deps.

I asked to get renovate/dependabot for the repo, so we ensure our stuff is
always compatible with the latest version (and should be compatible with
previous versions too but that's a different issue to tackle).