fix(Blob): Handle old module resolution types resolution (#377)
Before this commit, JavaScript projects, or TypeScript projects where the
moduleResolution setting was either undefined (= default to node10) or defined
but not part of node16 or nodenext would fail to properly autocomplete types for
import { upload } from '@vercel/blob/client'. As `/client` only exists when the
module resolution supports package.json `exports`. In all other situations you
need to configure `typesVersions` or create a folder to mimick the import.

This took me ages to solve, see https://stackoverflow.com/questions/58990498/package-json-exports-field-not-working-with-typescript