fix(blob): enforce content-type on fetch requests
Before this commit, we would not specify the type of body we were sending to the
customer routes handling token generation. On the app router example this is not
an issue because you have to express await req.json(). But on the pages router
situations, this is an issue because req.body is automatically filled/parsed
using the content-type.

We're fixing another place where we have this issue in our internal API
triggering upload completed.