fix(blob): error early when trying to use conflicting characters
Trying to upload files containing ?, # or // will always result in a weird
behavior where the pathname you sent won't be exactly the one in the resulting
url. Rather than relying on this I propose we just error early on and advise to
either not use them or encode them.

BREAKING CHANGE: Previously working characters will now throw an error.