Allow cacheLife to error
Next.js currently bundles proxy in a way that is indistinguishable from page code but in this context "use cache" isn't turned into a Cache Function so calls to `cacheLife` error. This will be fixed in Next.js but as an interrim fix we can just try/catch and suppress the error. This is safe for our context becasue we know we only call this from within "use cache" and we know that in any environment without "use cache" it should just fall back to normal behavior.